Added some css fixes and the files for MySql API
This should include the references to MySql.Data so that it
won't give you issues in the future.

Added a one line fix to css to make textboxes highlight
correctly on signup form.

For your email functionality, look to the assembly
System.Net.Mail, which includes an object called
SmtpClient. You can use this to send your email.

To add this, simply put using System.Net.Mail; at the top of your
code behind file.

In addition, it doesn't look like the conversion to asp.net
is the root cause of your css formatting issues. I could be wrong
but that is my impression, as your php pages seem to have
the same issue.